Fix for the issue of owl:Thing not being labelled as a Root in neo4j
An axiom subClassOf (someproperty some owl:Thing) could cause the owl:Thing
node to be created through the code path for "related nodes", which does
not set the Root label.

IMO the neo4j indexer is in need of refactoring to make it more explicit which
code paths are responsible for creating nodes, but meanwhile this minimal
fix simply ensures that all nodes are created before handling related nodes.